requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.13.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.1.0
Babel==2.9.1
billiard==4.2.1
cachetools==5.5.2
Cartopy==0.24.1
celery==5.4.0
certifi==2024.8.30
cfgv==3.4.0
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cron-descriptor==1.4.5
cycler==0.12.1
distlib==0.3.8
Django==4.2
django-celery-beat==2.7.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
fastapi==0.115.0
filelock==3.16.1
flet==0.25.1
flet-desktop==0.25.1
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
identify==2.6.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
kombu==5.4.2
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.10.0
multidict==6.1.0
netCDF4==1.7.2
nodeenv==1.9.1
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pre-commit==3.8.0
prompt_toolkit==3.0.50
propcache==0.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.1
pyproj==3.7.1
pyshp==2.3.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
repath==0.9.0
requests==2.32.3
ruff==0.6.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.35
sqlparse==0.5.3
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.30.6
vine==5.1.0
virtualenv==20.26.5
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
xarray==2025.1.2
yarl==1.18.3
